Optimised draw functions. Made option to parse decorations from file.
I finished making interface to parse decorations from file - to provide oportunity to to it for anyone without any difficulties. Then, i found one problem: call of "update" in drawBg function. It makes programm so much non-optimised. So i will try to fox it and make it as fast as i can, but tomorrow.